# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: ROSA Regional Frontend API
  description: Frontend API for ROSA HCP regional cluster management
  version: 0.0.1
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api/v0
    description: API v0
paths:
...
  /accounts/{accountId}/enable:
    post:
      summary: Enable account for authorization
      description: |
        Enables an AWS account for Cedar/AVP-based authorization.
        Creates a policy store in Amazon Verified Permissions.
        Requires privileged access.
      operationId: enableAccount
      tags:
        - Authorization
      parameters:
        - name: accountId
          in: path
          required: true
          description: AWS account ID to enable
          schema:
            type: string
            pattern: '^\d{12}$'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnableAccountRequest'
      responses:
        '201':
          description: Account enabled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - caller is not privileged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Account already enabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /accounts/{accountId}/disable:
    post:
      summary: Disable account authorization
      description: |
        Disables an AWS account from authorization.
        Deletes the policy store and all associated policies, groups, and attachments.
        Requires privileged access.
      operationId: disableAccount
      tags:
        - Authorization
      parameters:
        - name: accountId
          in: path
          required: true
          description: AWS account ID to disable
          schema:
            type: string
            pattern: '^\d{12}$'
      responses:
        '204':
          description: Account disabled successfully
        '403':
          description: Forbidden - caller is not privileged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Authorization - Policy Management
...
components:
  schemas:
...
    Error:
      type: object
      description: Error response
      required:
        - code
        - reason
      properties:
        id:
          type: string
          description: Error ID
        kind:
          type: string
          description: Resource kind
          example: Error
        href:
          type: string
          description: API href reference
        code:
          type: string
          description: Error code
        reason:
          type: string
          description: Human-readable error message
        operation_id:
          type: string
          description: Request operation ID for tracing
...
    EnableAccountRequest:
      type: object
      description: Request body for enabling an account
      properties:
        privileged:
          type: boolean
          description: If true, account bypasses all authorization checks
          default: false
    Account:
      type: object
      description: An enabled account
      required:
        - accountId
        - privileged
        - createdAt
      properties:
        accountId:
          type: string
          description: AWS account ID
          pattern: '^\d{12}$'
        policyStoreId:
          type: string
          description: AVP policy store ID (null if privileged)
        privileged:
          type: boolean
          description: If true, bypasses all authorization
        createdAt:
          type: string
          format: date-time
          description: When the account was enabled
        createdBy:
          type: string
          description: ARN of who enabled the account
...